Read and store HOTP counters in files
This commit allow the plugin to read and store HTOP counters in externals
files.

For each HOTP, we computed the SHA1 checksum of the secret key, which we
use as the filename. We store those files  in the
``/var/cache/openvpn/hotp-counters`` by default.

When adding a HOTP, the administrator is excepted to manually initalising the
counter file. This can be achivied with the following commands :

    PATH = /var/cache/openvpn/hotp-counters/"$(echo -n 'secretkey' | sha1sum | cut -f1 -d ' ')"
    echo -n 10 > $PATH